logic changes

+++ src/bookings/bookings.service.ts
@@ -1,18 +1,23 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
+import { BookingNotFoundException } from './exceptions/booking-not-found.exception';
 
 @Injectable()
-export class BookingsService { 
-    constructor(private prisma: PrismaService) {}
+export class BookingsService {
+    constructor(private prisma: PrismaService) { }
 
     async createBooking(tourId: string, email: string, name: string) {
-        return this.prisma.booking.create({
-            data: {
-                tourId,
-                email,
-                name
-            }
-        });
+        try {
+            return this.prisma.booking.create({
+                data: {
+                    tourId,
+                    email,
+                    name
+                }
+            });
+        } catch (error) {
+            throw new Error(String(error))
+        }
     }
 
     async getBookings() {
@@ -20,21 +25,47 @@ export class BookingsService {
     }
 
     async getBookingById(id: string) {
+        const booking = await this.prisma.booking.findUnique({ where: { id } })
+
+        if (!booking) {
+            throw new BookingNotFoundException(id);
+        }
+
         return this.prisma.booking.findUnique({
             where: { id }
         });
     }
 
     async deleteBooking(id: string) {
-        return this.prisma.booking.delete({
-            where: { id }
-        });
+        const booking = await this.prisma.booking.findUnique({ where: { id } })
+
+        if (!booking) {
+            throw new BookingNotFoundException(id);
+        }
+
+        try {
+            return this.prisma.booking.delete({
+                where: { id }
+            });
+        } catch (error) {
+            throw new Error(String(error))
+        }
     }
 
-    async updateBooking(id: string, data: { tourId?: string; email?: string; name?: string }) {
-        return this.prisma.booking.update({
-            where: { id },
-            data
-        });
+    async updateBooking(id: string, data: { email?: string; name?: string }) {
+        const booking = await this.prisma.booking.findUnique({ where: { id } });
+
+        if (!booking) {
+            throw new BookingNotFoundException(id);
+        }
+
+        try {
+            return this.prisma.booking.update({
+                where: { id },
+                data
+            });
+        } catch (error) {
+            throw new Error(String(error))
+        }
     }
 }
